+++ compiler/rustc_macros/src/lib.rs
@@ -127,6 +127,7 @@ decl_derive!(
         // struct attributes
         warning,
         error,
+        lint,
         note,
         help,
         // field attributes
@@ -138,6 +139,24 @@ decl_derive!(
         suggestion_short,
         suggestion_hidden,
         suggestion_verbose)] => diagnostics::session_diagnostic_derive
+);
+decl_derive!(
+    [LintDiagnostic, attributes(
+        // struct attributes
+        warning,
+        error,
+        lint,
+        note,
+        help,
+        // field attributes
+        skip_arg,
+        primary_span,
+        label,
+        subdiagnostic,
+        suggestion,
+        suggestion_short,
+        suggestion_hidden,
+        suggestion_verbose)] => diagnostics::lint_diagnostic_derive
 );
 decl_derive!(
     [SessionSubdiagnostic, attributes(
