+++ src/test/ui-fulldeps/session-diagnostic/diagnostic-derive.stderr
@@ -363,6 +363,28 @@ error: `#[label(...)]` is not a valid attribute
 LL |     #[label(typeck::label, foo("..."))]
    |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+error: only `#[error(..)]` and `#[warn(..)]` are supported
+  --> $DIR/diagnostic-derive.rs:540:1
+   |
+LL | / #[lint(typeck::ambiguous_lifetime_bound)]
+LL | |
+LL | | struct LintsBad {
+LL | | }
+   | |_^
+   |
+   = help: use the `#[error(...)]` attribute to create a error
+
+error: only `#[lint(..)]` is supported
+  --> $DIR/diagnostic-derive.rs:551:1
+   |
+LL | / #[error(typeck::ambiguous_lifetime_bound)]
+LL | |
+LL | | struct ErrorsBad {
+LL | | }
+   | |_^
+   |
+   = help: use the `#[lint(...)]` attribute to create a lint
+
 error: cannot find attribute `nonsense` in this scope
   --> $DIR/diagnostic-derive.rs:53:3
    |
@@ -395,7 +417,7 @@ LL |         arg: impl IntoDiagnosticArg,
    |                   ^^^^^^^^^^^^^^^^^ required by this bound in `DiagnosticBuilder::<'a, G>::set_arg`
    = note: this error originates in the derive macro `SessionDiagnostic` (in Nightly builds, run with -Z macro-backtrace for more info)
 
-error: aborting due to 46 previous errors
+error: aborting due to 48 previous errors
 
 Some errors have detailed explanations: E0277, E0425.
 For more information about an error, try `rustc --explain E0277`.
